+++ web/src/components/organisms/Settings/MyIntegrationDetails/index.tsx
@@ -1,4 +1,4 @@
-import MyIntegrationContent from "@reearth-cms/components/molecules/MyIntegrations/MyIntegrationContent";
+import MyIntegrationContent from "@reearth-cms/components/molecules/MyIntegrations/Content";
 
 const MyIntegrationDetails: React.FC = () => {
   return <MyIntegrationContent />;
